@@ -1,38 +1,30 @@
-// @ts-expect-error TS(2792): Cannot find module 'next/router'. Did you mean to ... Remove this comment to see the full error message
-import { useRouter } from 'next/router';
-// @ts-expect-error TS(2792): Cannot find module 'react'. Did you mean to set th... Remove this comment to see the full error message
-import React, { useContext, useEffect } from 'react';
 
-// @ts-expect-error TS(2792): Cannot find module '@chakra-ui/react'. Did you mea... Remove this comment to see the full error message
+import { useRouter } from 'next/router';
+import React, { useEffect } from 'react';
+
 import { Flex } from '@chakra-ui/react';
-// @ts-expect-error TS(2792): Cannot find module '@vercel/analytics'. Did you me... Remove this comment to see the full error message
 import { track } from '@vercel/analytics';
-// @ts-expect-error TS(2792): Cannot find module '@vercel/analytics/react'. Did ... Remove this comment to see the full error message
 import { Analytics } from '@vercel/analytics/react';
 
-// @ts-expect-error TS(2792): Cannot find module '../constants'. Did you mean to... Remove this comment to see the full error message
 import { SUPPORTED_NETWORKS } from '../constants';
-// @ts-expect-error TS(6142): Module './ConnectWeb3' was resolved to '/Users/moc... Remove this comment to see the full error message
 import { ConnectWeb3 } from './ConnectWeb3';
-// @ts-expect-error TS(6142): Module './Footer' was resolved to '/Users/moc/dev/... Remove this comment to see the full error message
 import { Footer } from './Footer';
-// @ts-expect-error TS(6142): Module './Header' was resolved to '/Users/moc/dev/... Remove this comment to see the full error message
 import { Header } from './Header';
+import { useWalletClient } from 'wagmi';
 
-// @ts-expect-error TS(6142): Module '../context/Web3Context' was resolved to '/... Remove this comment to see the full error message
-
-export function Layout({ children }: any) {
-  const { chain, account } = useWalletClient();
+export const Layout : React.FC<React.PropsWithChildren> = ({ children }) => {
+  const {data: walletClient} = useWalletClient();
+  const chainId = walletClient?.chain?.id;
+  const account = walletClient?.account;
 
   useEffect(() => {
-    track('ChainChanged', { chain });
-  }, [chain]);
+    track('ChainChanged', { chain:chainId ?? null});
+  }, [chainId]);
 
   const router = useRouter();
   const isOpenPath =
     router.pathname === '/' || router.pathname === '/contracts';
-  const isValid =
-    (account && SUPPORTED_NETWORKS.indexOf(chain) !== -1) || isOpenPath;
+  const isValid = (account && chainId && chainId in SUPPORTED_NETWORKS) || isOpenPath;
 
   return (
     <Flex
